Move GetBuilder to Utility Class
Remove GetBuilder from Repositories
